{"schema":"decodex/commit/1","summary":"Modularize live frame buffer","authority":"manual"}
